requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.25.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
slowapi>=0.1.9
python-multipart>=0.0.9
python-dotenv>=1.0.0
httpx>=0.27.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0
